Reorganize file structure for layout components
