+++ client/App.jsx
@@ -5,14 +5,14 @@ import {
   Route,
 } from 'react-router-dom';
 
-import Layout from './Layout.jsx';
+import Layout from './components/layout/Layout';
 import Login from './Login';
 import SignUp from './SignUp';
 import MapPage from './components/map/MapPage';
 import Homepage from './Homepage.jsx';
 import Dashboard from './components/dashboard/Dashboard';
 // import Add from './components/dashboard/AddItem.jsx';
-import ItemTable from './components/dashboard/ItemTable.jsx';
+// import ItemTable from './components/dashboard/ItemTable.jsx';
 
 import './styles/styles.css';
 
@@ -26,7 +26,7 @@ const App = () => {
         <Route exact path="/signup" element={<Layout><SignUp /></Layout>} />
         <Route exact path="/map" element={<Layout><MapPage /></Layout>} />
         <Route exact path="/dashboard" element={<Layout><Dashboard /></Layout>} />
-        <Route exact path="/test" element={<Layout><ItemTable /></Layout>} /> 
+        {/* <Route exact path="/test" element={<Layout><ItemTable /></Layout>} />  */}
       </Routes>
     </Router>
   );
