+++ client/components/layout/NavBar.jsx
@@ -7,14 +7,10 @@ import Typography from '@mui/material/Typography';
 import Menu from '@mui/material/Menu';
 import MenuIcon from '@mui/icons-material/Menu';
 import Container from '@mui/material/Container';
-import Avatar from '@mui/material/Avatar';
 import Button from '@mui/material/Button';
 import Tooltip from '@mui/material/Tooltip';
 import MenuItem from '@mui/material/MenuItem';
-import HomeIcon from '@mui/icons-material/Home';
 import { Link } from 'react-router-dom';
-
-
 
 const pages = ['Sign In', 'Sign Up', 'Chapters', 'Map'];
 const links = ['/signin', '/signup', '/chapters', '/map'];
